Refactor imports; add test tsconfig
Adjust module imports and TypeScript configs, and small housekeeping.

- Update several imports to point to concrete module paths (e.g. types/King, types/PieceType, src/Game, src/Board, src/pieces, src/positions) to fix resolution.
- Simplify game-engine/tsconfig.json for production (remove jest/node types, restrict include to ./src) and add game-engine/tsconfig.test.json that extends it and enables jest/node types for tests (noEmit).
- Add debug.log (runtime error log) to repo.
- Remove an unnecessary ESLint disable comment in server lobby.service after removing the unused variable.
