+++ debug.log
@@ -1,0 +1,1 @@
+[0331/204249.670:ERROR:third_party\crashpad\crashpad\util\win\registration_protocol_win.cc:108] CreateFile: Nie można odnaleźć określonego pliku. (0x2)
