+++ game-engine/tsconfig.json
@@ -1,18 +1,19 @@
 {
-  "compilerOptions": {
-    "target": "es2016",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
-    "module": "CommonJS",                                /* Specify what module code is generated. */   /* Allow 'import x from y' when a module doesn't have a default export. */
-    "esModuleInterop": true,                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
-    "forceConsistentCasingInFileNames": true,            /* Ensure that casing is correct in imports. */
-    "strict": true,                                      /* Enable all strict type-checking options. */
-    "skipLibCheck": true,                                 /* Skip type checking all .d.ts files. */
-    "outDir": "build",
-    "declaration": true,
-    "types": ["jest", "node"]
-  },
-  "include": [
-    "./src/**/*.ts",
-    "./e2e-tests/**/*.ts",
-    "./test-utils/**/*.ts"
-  ],
+    "compilerOptions": {
+        "target": "es2016",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
+        "module": "CommonJS",                                /* Specify what module code is generated. */
+        "esModuleInterop": true,                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
+        "forceConsistentCasingInFileNames": true,            /* Ensure that casing is correct in imports. */
+        "strict": true,                                      /* Enable all strict type-checking options. */
+        "skipLibCheck": true,                                /* Skip type checking all .d.ts files. */
+
+        "outDir": "build",
+        "declaration": true,
+
+        "types": []
+    },
+
+    "include": [
+        "./src/**/*.ts"
+    ]
 }
