+++ game-engine/tsconfig.test.json
@@ -1,0 +1,14 @@
+{
+    "extends": "./tsconfig.json",
+
+    "compilerOptions": {
+        "types": ["jest", "node"],
+        "noEmit": true
+    },
+
+    "include": [
+        "./src/**/*.ts",
+        "./e2e-tests/**/*.ts",
+        "./test-utils/**/*.ts"
+    ]
+}
